Add inline declarations to translation units and make variables static
